conclusion: didn't need to explicitly maximize
Spacemacs has a configuration setting for that. I was just being dumb
because the semantic layer was erroring out, causing my config to not
finish loading. Thus the fullscreen wasn't working (among other things).
The semantic layer seems broken; will not disable it for now so it will
hopefully resolve itself upstream.